+++ CHANGELOG.md
@@ -49,6 +49,8 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Resource management REST API documentation (@dharezlak)
 - Service aliases (@jmeizner)
 - Issue and merge request templates (@mkasztelnik)
+- Users can set Access Methods for new and existing Services (@Nuanda)
+- There are global Access Methods which work for all Services (@Nuanda)
 
 ### Changed
 - Upgrade to Rails 5 (@mkasztelnik, @Nuanda)